requirements.txt
black==24.8.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.5
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.9.1.post1
mdurl==0.1.2
mkl-service==2.4.0
mypy-extensions==1.0.0
opencv-python==4.10.0.84
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
protobuf==5.27.3
pyparsing==3.1.2
pytz==2024.1
pywin32==306
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentry-sdk==2.13.0
setproctitle==1.3.3
smmap==5.0.1
threadpoolctl==3.5.0
timm==1.0.8
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.5
tzdata==2024.1
wandb==0.17.7
